[oneDPL][ranges][sycl] + a fix for header inclusion order (<ranges> / <execution>); + a couple of tests to cover that situation.
