+++ include/oneapi/dpl/pstl/hetero/algorithm_ranges_impl_hetero.h
@@ -18,6 +18,7 @@
 
 #include "../algorithm_fwd.h"
 #include "../parallel_backend.h"
+#include "utils_hetero.h"
 
 #if _ONEDPL_BACKEND_SYCL
 #    include "dpcpp/utils_ranges_sycl.h"
