+++ include/oneapi/dpl/pstl/hetero/dpcpp/parallel_backend_sycl_utils.h
@@ -230,34 +230,6 @@ inline void __print_device_debug_info(_Policy, size_t = 0, size_t = 0)
 }
 #endif
 
-// struct for checking if iterator is heterogeneous or not
-template <typename Iter, typename Void = void> // for non-heterogeneous iterators
-struct is_hetero_iterator : ::std::false_type
-{
-};
-
-template <typename Iter> // for heterogeneous iterators
-struct is_hetero_iterator<Iter, typename ::std::enable_if<Iter::is_hetero::value, void>::type> : ::std::true_type
-{
-};
-// struct for checking if iterator should be passed directly to device or not
-template <typename Iter, typename Void = void> // for iterators that should not be passed directly
-struct is_passed_directly : ::std::false_type
-{
-};
-
-template <typename Iter> // for iterators defined as direct pass
-struct is_passed_directly<Iter, typename ::std::enable_if<Iter::is_passed_directly::value, void>::type>
-    : ::std::true_type
-{
-};
-
-template <typename Iter> // for pointers to objects on device
-struct is_passed_directly<Iter, typename ::std::enable_if<::std::is_pointer<Iter>::value, void>::type>
-    : ::std::true_type
-{
-};
-
 //-----------------------------------------------------------------------
 // type traits for comparators
 //-----------------------------------------------------------------------
